remove resource service
